feat: mobile chat view
