--- projects/app/src/app/(chat)/chat/layout.tsx
@@ -1,7 +1,0 @@
-export default async function Layout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
