+++ projects/app/src/components/ai-elements/actions.tsx
@@ -35,7 +35,7 @@ export const Action = ({
   const button = (
     <Button
       className={cn(
-        'text-muted-foreground hover:text-foreground size-9 p-1.5',
+        'text-muted-foreground hover:text-foreground size-12 p-2 md:size-9 md:p-1.5',
         className,
       )}
       size={size}
